let the caller handle ConcurrentModificationException in transactions
If several instances try to write the same entity in a transaction it can cause contention on that entity. In case of contention the current code retries the operation endlessly without any wait period which makes the contention even worse. This makes it impossible for any transaction to successfully write the entity which leads to an endless loop of transaction retries.

We must use exponential backoff between transaction retries to ease the load on datastore. will happen in this case.